pyproject.toml
[project]
name = "aula17"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.1.3",
    "langchain-anthropic>=1.2.0",
    "langchain-google-genai>=4.0.0",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "numexpr>=2.14.1",
    "python-dotenv>=1.2.1",
]
